Style the first div element with the class selector .recipe
Add a border of 2px and make it solid black. Give the box a background
color of white and make its position property relative. Change the
boxes height to 825px and its width to 785px. Give the box a padding
property of 35px and set its left and right margin properties to auto.