Edit lemon_pepper_chicken.html and styles.css
Give the div element with the class name "box" a second class name of
"r1" within the lemon_pepper_chicken.html file. Then remove the height
and width properties in the .box class selector. Add a new class
selecor of .box.r1 and add the height and width properties that were
removed from the prievious selctor into the new selector.