Style the image in fruity_pebble_krispies.html within styles.css
Give the image an id name of "krispies" then reajust its height and
width to 395px by 340px.
In styles.css, add an id selector of "#krispies" then give the image a
border of 1px and make it solid black. Add a position property and set
it to absolute then add a Top property of 35px.